Move the recent invoices screen out in the home screen
