+++ Navigation.js
@@ -12,9 +12,9 @@ import Currency from './screens/settings/Currency';
 import Settings from './screens/settings/Settings';
 import ShopName from './screens/settings/ShopName';
 import { TouchableOpacity } from 'react-native-gesture-handler';
-import RecentInvoices from './screens/settings/recentInvoices/RecentInvoices';
-import InvoiceDetail from './screens/settings/recentInvoices/InvoiceDetail';
-import ScanInvoice from './screens/settings/recentInvoices/ScanInvoice';
+import RecentInvoices from './screens/recentInvoices/RecentInvoices';
+import InvoiceDetail from './screens/recentInvoices/InvoiceDetail';
+import ScanInvoice from './screens/recentInvoices/ScanInvoice';
 
 
 const boltPosLogo = require('./img/bolt-card-pos.png');
@@ -42,7 +42,6 @@ const SettingsRoot = () => {
       screenOptions={{headerHideShadow: true, headerShown:false}}
       initialRouteName="Settings">
       <SettingsStack.Screen name="Settings" component={Settings} />
-      <SettingsStack.Screen name="Recent Invoices" component={InvoiceRoot} />
       <SettingsStack.Screen name="Connect" component={ConnectToHub} />
       <SettingsStack.Screen name="Currency" component={Currency} />
       <SettingsStack.Screen name="Shop Name" component={ShopName} />
@@ -96,6 +95,11 @@ const HomeRoot = () => {
         component={SettingsRoot}
         options={{...NavigationDefaultOptions, headerStyle: {backgroundColor: isDarkMode ? '#666' : '#fff'}}}
       />
+      <HomeStack.Screen 
+        name="Recent Invoices" 
+        component={InvoiceRoot}
+        options={{...NavigationDefaultOptions, headerStyle: {backgroundColor: isDarkMode ? '#666' : '#fff'}}}
+       />
     </HomeStack.Navigator>
   );
 };
